add deleteOrder and updateOrder

+++ src/main/java/ro/myclass/onlineStoreapi/repo/OrderRepo.java
@@ -13,10 +13,13 @@ import java.util.Optional;
 public interface OrderRepo extends JpaRepository<Order,Long> {
 
     @Query("select o from Order o where o.customer.id = ?1")
-    List<Order> getOrderByCustomerId(int customerId);
+    List<Order> getOrderByCustomerId(long customerId);
 
    @Query("select o from Order o where o.customer.email = ?1 and o.orderDate = ?2")
    Optional<Order> getOrderByCustomerEmailAndOrderDate(String email, Date order);
+
+
+
